Reintroduce explicit public class GraphQlQuery to be used with AEMHeadlessClient
#18
